requirements.txt
Django==1.8.1
Pillow==2.8.1
argparse==1.2.1
psycopg2==2.6
wsgiref==0.1.2
